+++ src/plumbing.rs
@@ -5,10 +5,10 @@ use crate::MenuItem;
 
 /// Menu-related extensions for object chain elements
 pub trait MenuItemCollection<R, S> {
-    fn bounds_of(&self, nth: u32) -> Rectangle;
-    fn title_of(&self, nth: u32) -> &str;
-    fn details_of(&self, nth: u32) -> &str;
-    fn interact_with(&mut self, nth: u32) -> R;
+    fn bounds_of(&self, nth: usize) -> Rectangle;
+    fn title_of(&self, nth: usize) -> &str;
+    fn details_of(&self, nth: usize) -> &str;
+    fn interact_with(&mut self, nth: usize) -> R;
     fn count(&self) -> usize;
 }
 
@@ -17,22 +17,22 @@ impl<I, R, S> MenuItemCollection<R, S> for I
 where
     I: MenuItem<R, S> + View + crate::Marker,
 {
-    fn bounds_of(&self, nth: u32) -> Rectangle {
+    fn bounds_of(&self, nth: usize) -> Rectangle {
         debug_assert!(nth == 0);
         self.bounds()
     }
 
-    fn interact_with(&mut self, nth: u32) -> R {
+    fn interact_with(&mut self, nth: usize) -> R {
         debug_assert!(nth == 0);
         self.interact()
     }
 
-    fn title_of(&self, nth: u32) -> &str {
+    fn title_of(&self, nth: usize) -> &str {
         debug_assert!(nth == 0);
         self.title()
     }
 
-    fn details_of(&self, nth: u32) -> &str {
+    fn details_of(&self, nth: usize) -> &str {
         debug_assert!(nth == 0);
         self.details()
     }
@@ -46,19 +46,19 @@ impl<I, R, S> MenuItemCollection<R, S> for Chain<I>
 where
     I: MenuItemCollection<R, S>,
 {
-    fn bounds_of(&self, nth: u32) -> Rectangle {
+    fn bounds_of(&self, nth: usize) -> Rectangle {
         self.object.bounds_of(nth)
     }
 
-    fn interact_with(&mut self, nth: u32) -> R {
+    fn interact_with(&mut self, nth: usize) -> R {
         self.object.interact_with(nth)
     }
 
-    fn title_of(&self, nth: u32) -> &str {
+    fn title_of(&self, nth: usize) -> &str {
         self.object.title_of(nth)
     }
 
-    fn details_of(&self, nth: u32) -> &str {
+    fn details_of(&self, nth: usize) -> &str {
         self.object.details_of(nth)
     }
 
@@ -72,8 +72,8 @@ where
     I: MenuItemCollection<R, S>,
     LE: MenuItemCollection<R, S> + ChainElement,
 {
-    fn bounds_of(&self, nth: u32) -> Rectangle {
-        let count = self.object.count() as u32;
+    fn bounds_of(&self, nth: usize) -> Rectangle {
+        let count = self.object.count();
         if nth < count {
             self.object.bounds_of(nth)
         } else {
@@ -81,8 +81,8 @@ where
         }
     }
 
-    fn interact_with(&mut self, nth: u32) -> R {
-        let count = self.object.count() as u32;
+    fn interact_with(&mut self, nth: usize) -> R {
+        let count = self.object.count();
         if nth < count {
             self.object.interact_with(nth)
         } else {
@@ -90,8 +90,8 @@ where
         }
     }
 
-    fn title_of(&self, nth: u32) -> &str {
-        let count = self.object.count() as u32;
+    fn title_of(&self, nth: usize) -> &str {
+        let count = self.object.count();
         if nth < count {
             self.object.title_of(nth)
         } else {
@@ -99,8 +99,8 @@ where
         }
     }
 
-    fn details_of(&self, nth: u32) -> &str {
-        let count = self.object.count() as u32;
+    fn details_of(&self, nth: usize) -> &str {
+        let count = self.object.count();
         if nth < count {
             self.object.details_of(nth)
         } else {
